Begin to split metrics.hpp header into multiple files.
